requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
slack_sdk==3.27.0
